protected delete post with authorization

+++ backend/routes/posts.js
@@ -61,6 +61,7 @@ router.put("/:id", checkAuth, multer({storage}).single("image"), (req, res, next
   });
   Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
     if (result.nModified > 0) {
+      res.status(200).json({ message: "Updated successfully!" });
     } else {
       res.status(403).json({ message: "Forbidden!" });
     }
@@ -100,9 +101,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: "Post deleted successfully!" });
+    } else {
+      res.status(403).json({ message: "Forbidden!" });
+    }
   });
 });
